Move location description rule to new home
